+++ src/components/Header.astro
@@ -1,5 +1,5 @@
 ---
-import catalog from '@eventcatalog';
+import catalog from '@utils/eventcatalog-config/catalog';
 import Search from '@components/Search.astro';
 import { buildUrl } from '@utils/url-builder';
