+++ src/layouts/DiscoverLayout.astro
@@ -3,10 +3,10 @@ import { Table } from '@components/Tables/Table';
 import { QueueListIcon, RectangleGroupIcon, BoltIcon, ChatBubbleLeftIcon } from '@heroicons/react/24/outline';
 import ServerIcon from '@heroicons/react/24/outline/ServerIcon';
 import { getCommands } from '@utils/commands';
-import { getDomains } from '@utils/domains/domains';
-import { getFlows } from '@utils/flows/flows';
+import { getDomains } from '@utils/collections/domains';
+import { getFlows } from '@utils/collections/flows';
 import { getEvents } from '@utils/events';
-import { getServices } from '@utils/services/services';
+import { getServices } from '@utils/collections/services';
 import { buildUrl } from '@utils/url-builder';
 import { getQueries } from '@utils/queries';
 import { MagnifyingGlassIcon } from '@heroicons/react/20/solid';
