+++ src/layouts/VerticalSideBarLayout.astro
@@ -10,13 +10,13 @@ import Header from '../components/Header.astro';
 import SEO from '../components/Seo.astro';
 
 import { getCommands } from '@utils/commands';
-import { getDomains } from '@utils/domains/domains';
+import { getDomains } from '@utils/collections/domains';
 import { getEvents } from '@utils/events';
-import { getServices } from '@utils/services/services';
-import { getFlows } from '@utils/flows/flows';
+import { getServices } from '@utils/collections/services';
+import { getFlows } from '@utils/collections/flows';
 import { getTeams } from '@utils/teams';
 import { getUsers } from '@utils/users';
-import config, { type CatalogConfig } from '@eventcatalog';
+import config, { type CatalogConfig } from '@utils/eventcatalog-config/catalog';
 import { buildUrl } from '@utils/url-builder';
 import { getQueries } from '@utils/queries';
 import { getChannels } from '@utils/channels';
