+++ src/pages/docs/[type]/[id]/[version]/index.astro
@@ -15,8 +15,8 @@ import { QueueListIcon, RectangleGroupIcon, ServerIcon, BoltIcon, ChatBubbleLeft
 import type { PageTypes } from '@types';
 
 import { buildUrl } from '@utils/url-builder';
-import { getFlows } from '@utils/flows/flows';
-import { pageDataLoader } from '@utils/pages/pages';
+import { getFlows } from '@utils/collections/flows';
+import { pageDataLoader } from '@utils/page-loaders/page-data-loader';
 import { ViewTransitions } from 'astro:transitions';
 import { ArrowsRightLeftIcon } from '@heroicons/react/20/solid';
