+++ src/pages/docs/[type]/[id]/[version]/spec/index.astro
@@ -7,7 +7,7 @@ import type { CollectionTypes, PageTypes } from '@types';
 import PlainPage from '@layouts/PlainPage.astro';
 import { DocumentMinusIcon } from '@heroicons/react/24/outline';
 import { buildUrl } from '@utils/url-builder';
-import { pageDataLoader } from '@utils/pages/pages';
+import { pageDataLoader } from '@utils/page-loaders/page-data-loader';
 import VerticalSideBarLayout from '@layouts/VerticalSideBarLayout.astro';
 import './styles.css';
