+++ src/pages/index.astro
@@ -10,9 +10,9 @@ import {
 } from '@heroicons/react/24/outline';
 
 import { getMessages } from '@utils/messages';
-import { getDomains } from '@utils/domains/domains';
-import { getServices } from '@utils/services/services';
-import { getFlows } from '@utils/flows/flows';
+import { getDomains } from '@utils/collections/domains';
+import { getServices } from '@utils/collections/services';
+import { getFlows } from '@utils/collections/flows';
 import DiscoverInsight from '@components/DiscoverInsight.astro';
 import VerticalSideBarLayout from '@layouts/VerticalSideBarLayout.astro';
 import { BookOpen, BookOpenText, TableProperties, Workflow } from 'lucide-react';
