+++ src/pages/visualiser/[type]/[id]/index.astro
@@ -3,8 +3,8 @@ import Seo from '@components/Seo.astro';
 import { buildUrl } from '@utils/url-builder';
 import { getEvents } from '@utils/events';
 import { getCommands } from '@utils/commands';
-import { getServices } from '@utils/services/services';
-import { getDomains } from '@utils/domains/domains';
+import { getServices } from '@utils/collections/services';
+import { getDomains } from '@utils/collections/domains';
 import type { CollectionEntry } from 'astro:content';
 import type { CollectionTypes } from '@types';
