+++ src/utils/__tests__/domains/node-graph.spec.ts
@@ -1,7 +1,7 @@
 import type { ContentCollectionKey } from 'astro:content';
 import { expect, describe, it, vi } from 'vitest';
 import { mockDomains, mockServices, mockEvents, mockCommands } from './mocks';
-import { getNodesAndEdges } from '../../domains/node-graph';
+import { getNodesAndEdges } from '@utils/node-graphs/domains-node-graph';
 import { MarkerType } from 'reactflow';
 
 vi.mock('astro:content', async (importOriginal) => {
