+++ src/utils/collections/domains.ts
