+++ src/utils/page-loaders/page-data-loader.ts
@@ -1,9 +1,9 @@
 import type { CollectionTypes, PageTypes } from '@types';
 import { getChannels } from '@utils/channels';
-import { getDomains } from '@utils/domains/domains';
+import { getDomains } from '@utils/collections/domains';
 import { getCommands, getEvents } from '@utils/messages';
 import { getQueries } from '@utils/queries';
-import { getServices } from '@utils/services/services';
+import { getServices } from '@utils/collections/services';
 import type { CollectionEntry } from 'astro:content';
 
 export const pageDataLoader: Record<PageTypes, () => Promise<CollectionEntry<CollectionTypes>[]>> = {
