+++ tsconfig.json
@@ -5,7 +5,7 @@
     "baseUrl": ".",
     "paths": {
       "@config": ["./eventcatalog.config.js"],
-      "@eventcatalog": ["src/utils/config/catalog.ts"],
+      "@eventcatalog": ["src/utils/eventcatalog-config/catalog.ts"],
       "@components/*": ["src/components/*"],
       "@catalog/components/*": ["src/custom-defined-components/*"],
       "@types": ["src/types/index.ts"],
